FULLY WORKING

+++ onnxruntime/core/providers/js/js_export.cc
@@ -5,9 +5,9 @@
 
 #include "core/framework/op_kernel.h"
 
-const void* JsepOutput(void* context, int index, void* data) {
-  uint32_t* data_offset = reinterpret_cast<uint32_t*>(data);
-  uint32_t dim = *data_offset++;
+const void* JsepOutput(void* context, size_t index, void* data) {
+  size_t* data_offset = reinterpret_cast<size_t*>(data);
+  size_t dim = *data_offset++;
   size_t dim_size = static_cast<size_t>(dim);
   std::vector<int64_t> dims;
   dims.reserve(dim_size);
